npus-aiml-agentic-automation/crewai/requirements.txt
crewai 
crewai-tools[mcp] 
databricks-sdk 
fastapi 
uvicorn 
python-dotenv
langchain
langchain-chroma